@@ -24,7 +24,7 @@ class AwRestProblemExtension extends Extension
     /**
      * {@inheritdoc}
      */
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $configuration = new \Alterway\Bundle\RestProblemBundle\DependencyInjection\Configuration();
         $config = $this->processConfiguration($configuration, $configs);
